requirements.txt
torch>=2.0.1
torchaudio>=2.0.2
numpy>=1.20.0
scipy>=1.7.0
librosa>=0.9.0
phonemizer>=3.0.0
matplotlib>=3.5.0
tensorboard>=2.8.0
PyYAML>=6.0
tqdm>=4.64.0
soundfile>=0.10.3
Unidecode>=1.3.4
inflect>=6.0.0
transformers>=4.25.0
einops>=0.6.0
pandas>=1.5.0
wandb>=0.13.0
fastapi>=0.95.0
uvicorn>=0.21.0 